requirements.txt
flask>=3.0
flask-compress>=1.14
gunicorn>=22.0,<25.0
werkzeug>=3.0
supabase>=2.0
requests>=2.31
biopython>=1.81
numpy>=1.24
stripe>=9.0,<13.0
python-dotenv>=1.0
prometheus-client>=0.20

# Epitope Scout analysis pipeline (merged from epitope-scout/requirements.txt)
freesasa==2.2.1
scipy>=1.11
anthropic>=0.25
